Move migrate commands
