+++ .gitignore
@@ -10,3 +10,4 @@ node_modules/
 
 .virtualgo
 boilerplate/lyft/end2end/tmp
+dist
