+++ cmd/entrypoints/migrate.go
@@ -2,13 +2,8 @@ package entrypoints
 
 import (
 	"context"
+	"github.com/flyteorg/flyteadmin/pkg/server"
 
-	"github.com/flyteorg/flyteadmin/pkg/repositories"
-
-	"github.com/flyteorg/flyteadmin/pkg/repositories/config"
-	"github.com/flyteorg/flyteadmin/pkg/runtime"
-	"github.com/flyteorg/flytestdlib/logger"
-	"github.com/go-gormigrate/gormigrate/v2"
 	"github.com/spf13/cobra"
 	_ "gorm.io/driver/postgres" // Required to import database driver.
 )
@@ -22,36 +17,9 @@ var parentMigrateCmd = &cobra.Command{
 var migrateCmd = &cobra.Command{
 	Use:   "run",
 	Short: "This command will run all the migrations for the database",
-	Run: func(cmd *cobra.Command, args []string) {
+	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := context.Background()
-		configuration := runtime.NewConfigurationProvider()
-		databaseConfig := configuration.ApplicationConfiguration().GetDbConfig()
-		logConfig := logger.GetConfig()
-
-		db, err := repositories.GetDB(ctx, databaseConfig, logConfig)
-		if err != nil {
-			logger.Fatal(ctx, err)
-		}
-
-		sqlDB, err := db.DB()
-		if err != nil {
-			logger.Fatal(ctx, err)
-		}
-
-		defer func(deferCtx context.Context) {
-			if err = sqlDB.Close(); err != nil {
-				logger.Fatal(deferCtx, err)
-			}
-		}(ctx)
-
-		if err = sqlDB.Ping(); err != nil {
-			logger.Fatal(ctx, err)
-		}
-		m := gormigrate.New(db, gormigrate.DefaultOptions, config.Migrations)
-		if err = m.Migrate(); err != nil {
-			logger.Fatalf(ctx, "Could not migrate: %v", err)
-		}
-		logger.Infof(ctx, "Migration ran successfully")
+		return server.Migrate(ctx)
 	},
 }
 
@@ -59,36 +27,9 @@ var migrateCmd = &cobra.Command{
 var rollbackCmd = &cobra.Command{
 	Use:   "rollback",
 	Short: "This command will rollback one migration",
-	Run: func(cmd *cobra.Command, args []string) {
+	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := context.Background()
-		configuration := runtime.NewConfigurationProvider()
-		databaseConfig := configuration.ApplicationConfiguration().GetDbConfig()
-		logConfig := logger.GetConfig()
-
-		db, err := repositories.GetDB(ctx, databaseConfig, logConfig)
-		if err != nil {
-			logger.Fatal(ctx, err)
-		}
-		sqlDB, err := db.DB()
-		if err != nil {
-			logger.Fatal(ctx, err)
-		}
-		defer func(deferCtx context.Context) {
-			if err = sqlDB.Close(); err != nil {
-				logger.Fatal(deferCtx, err)
-			}
-		}(ctx)
-
-		if err = sqlDB.Ping(); err != nil {
-			logger.Fatal(ctx, err)
-		}
-
-		m := gormigrate.New(db, gormigrate.DefaultOptions, config.Migrations)
-		err = m.RollbackLast()
-		if err != nil {
-			logger.Fatalf(ctx, "Could not rollback latest migration: %v", err)
-		}
-		logger.Infof(ctx, "Rolled back one migration successfully")
+		return server.Rollback(ctx)
 	},
 }
 
@@ -96,36 +37,9 @@ var rollbackCmd = &cobra.Command{
 var seedProjectsCmd = &cobra.Command{
 	Use:   "seed-projects",
 	Short: "Seed projects in the database.",
-	Run: func(cmd *cobra.Command, args []string) {
+	RunE: func(cmd *cobra.Command, args []string) error {
 		ctx := context.Background()
-		configuration := runtime.NewConfigurationProvider()
-		databaseConfig := configuration.ApplicationConfiguration().GetDbConfig()
-		logConfig := logger.GetConfig()
-
-		db, err := repositories.GetDB(ctx, databaseConfig, logConfig)
-		if err != nil {
-			logger.Fatal(ctx, err)
-		}
-
-		sqlDB, err := db.DB()
-		if err != nil {
-			logger.Fatal(ctx, err)
-		}
-
-		defer func(deferCtx context.Context) {
-			if err = sqlDB.Close(); err != nil {
-				logger.Fatal(deferCtx, err)
-			}
-		}(ctx)
-
-		if err = sqlDB.Ping(); err != nil {
-			logger.Fatal(ctx, err)
-		}
-
-		if err = config.SeedProjects(db, args); err != nil {
-			logger.Fatalf(ctx, "Could not add projects to database with err: %v", err)
-		}
-		logger.Infof(ctx, "Successfully added projects to database")
+		return server.SeedProjects(ctx, args)
 	},
 }
